% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{empty_on_cran}
\alias{empty_on_cran}
\title{Reduce a check to the empty string on CRAN}
\usage{
empty_on_cran(string)
}
\arguments{
\item{string}{The string to test.}
}
\description{
In testing, we don't want to rely on exact error messages or stdout on CRAN
for third-party packages, since this would complicate the release process for
dependencies of \{precommit\}.
}
\keyword{internal}
